+++ src/CodeGenerator/src/Generator/ResponseParser/JsonRpcParser.php
@@ -16,7 +16,7 @@ class JsonRpcParser extends RestJsonParser
     protected function parseResponseTimestamp(Shape $shape, string $input, bool $required): string
     {
         if ($required) {
-            $body = '/** @var \DateTimeImmutable $d */ $d = \DateTimeImmutable::createFromFormat(\'U.u\', \sprintf(\'%.6F\', INPUT))';
+            $body = $this->rejectUnparsableTimestamp('\DateTimeImmutable::createFromFormat(\'U.u\', \sprintf(\'%.6F\', INPUT))');
         } else {
             $body = '(isset(INPUT) && ($d = \DateTimeImmutable::createFromFormat(\'U.u\', \sprintf(\'%.6F\', INPUT)))) ? $d : null';
         }